moved config fixtures to a conftest.py file
